Remove bundler-cache from sample config
This was added as a recommended part of the sample config, but it's not clear why it was added. That config runs `bundle install` and caches the installed gems. Since balto-brakeman does the work of only installing the minimum gems necessary for it to run (i.e. brakeman), also having a project do a full `bundle install` seems like extra work that's not needed, especially for larger apps.  A number of apps use brakeman without this option, and everything seems to work just fine.